First Update
This will be a repo for my 3D game engine created in C++

I am using the following frameworks/libraries:
-OpenGL (Graphics API)
-GLFW (Window/Input Callback Library)
-GLEW (Allows me to use Modern OpenGL)
-SOIL (Allows me to load images)
-GLM (Allows me to use maths involved with OpenGL)

I originally was writing my own maths library, but it was a lot of work,
when I could just use a great library!

So far the engine is pretty barebones, but it will be worked on over the
next few years so I can learn how game engines work!